+++ admin/src/pages/chats/ChatRoomPage.tsx
@@ -5,36 +5,25 @@ import {
   MainContainer,
   PageContainer,
 } from "../../atoms/layouts/PageContainer";
-import { getChatComments, patchChatEnd } from "../../services/CustomerService";
+import { patchChatEnd } from "../../services/CustomerService";
 import { setRoomId } from "../../redux/createSlice/customerInfoSlice";
 import React, { useEffect, useRef, useState } from "react";
-import { useMutation, useQuery } from "@tanstack/react-query";
-import { useToken } from "../../hooks/useToken";
+import { useMutation } from "@tanstack/react-query";
 import SockJS from "sockjs-client";
 import { Stomp } from "@stomp/stompjs";
-import Loading from "../../components/loading/Loading";
 import styled from "styled-components";
 import { queryClient } from "../..";
 import { PageTitle } from "../../styles/PageTitle";
 import tokens from "../../styles/tokens.json";
 import { BodyTextTypo } from "../../atoms/typographys/Typographys";
-import axios from "axios";
-import { errorResponseDataType } from "../../types/axiosErrorType";
 import { RegularButton } from "../../atoms/buttons/Buttons";
 import { ReactComponent as SendArrow } from "../../assets/images/icons/Send.svg";
 import { DarkMode } from "../../types/reportDataType";
-
-type ChatContents = {
-  message: string;
-  roomId: string;
-  sendDate: string;
-  sender: string;
-};
+import ChatList from "../../components/CustomerService/ChatList";
 
 const ChatRoomPage: React.FC = () => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
-  const refreshToken = useToken();
   const stompClient = useRef<any>(null);
   const scrollRef = useRef<HTMLUListElement>(null);
   const messageEndRef = useRef<HTMLDivElement>(null);
@@ -43,48 +32,14 @@ const ChatRoomPage: React.FC = () => {
   const accessToken = useSelector(
     (state: RootState) => state.loginInfo.accessToken
   );
-  const [isPage, setPage] = useState<number>(1);
   const [isMsg, setMsg] = useState<string>("");
+  const [isArrive, setArrive] = useState({});
+  const [isChatStart, setChatStart] = useState(false);
 
   // token
   const headers = {
     Authorization: accessToken.authorization,
   };
-
-  // 채팅방 대화 내용 조회
-  const {
-    data: ChatContents,
-    isLoading,
-    isFetching,
-    error,
-    isPreviousData,
-  } = useQuery({
-    queryKey: ["ChatComments"],
-    queryFn: async () => {
-      try {
-        const response = await getChatComments(
-          accessToken.authorization,
-          roomId,
-          isPage
-        );
-        console.log(response);
-        return response;
-      } catch (err) {
-        if (axios.isAxiosError<errorResponseDataType, any>(err)) {
-          if (err.response?.data.message === "만료된 토큰입니다.") {
-            refreshToken();
-          }
-        } else {
-          console.log(err);
-        }
-      }
-    },
-    keepPreviousData: true,
-    staleTime: 1000 * 60 * 5,
-    cacheTime: 1000 * 60 * 30,
-    retry: 3,
-    retryDelay: 1000,
-  });
 
   // STOMP 연결 설정
   const stompConnect = () => {
@@ -99,7 +54,7 @@ const ChatRoomPage: React.FC = () => {
           `/sub/chat/room/${roomId}`,
           (data: any) => {
             const newMessage = JSON.parse(data.body);
-            // console.log(newMessage);
+            setArrive(newMessage);
           },
           headers
         );
@@ -113,9 +68,7 @@ const ChatRoomPage: React.FC = () => {
   // STOMP 연결 해제
   const stompDisConnect = () => {
     try {
-      stompClient.current.disconnect(() => {
-        stompClient.current.unsubscribe(`/sub/chat/room/${roomId}`);
-      });
+      stompClient.current.disconnect(() => {});
       console.log("WebSocket 연결이 끊겼습니다. (stomp)");
       mutate();
     } catch (err) {
@@ -126,9 +79,6 @@ const ChatRoomPage: React.FC = () => {
   // message 보내기
   const handleSendMessage = async () => {
     if (isMsg === "") return;
-    if (messageEndRef.current) {
-      messageEndRef.current.scrollIntoView({ behavior: "smooth" });
-    }
     const newMessage = {
       roomId: roomId,
       message: isMsg,
@@ -176,83 +126,57 @@ const ChatRoomPage: React.FC = () => {
     }
   }, []);
 
-  useEffect(() => {
-    if (scrollRef.current) {
-      scrollRef.current.scrollTop = scrollRef.current.scrollHeight;
-    }
-  }, []);
-
   return (
     <PageContainer isDark={isDark}>
       <MainContainer isDark={isDark}>
         <PageTitle isDark={isDark}>고객센터</PageTitle>
 
-        {isLoading ? (
-          <Loading />
-        ) : (
-          <ChatContainer isDark={isDark}>
-            <ChatHeader isDark={isDark}>
-              <QuestionEndBtn
-                isDark={isDark}
-                onClick={() => {
-                  patchChatEnd(accessToken.authorization, roomId);
-                }}
-              >
-                상담 종료하기
-              </QuestionEndBtn>
-              <RoomID isDark={isDark}>ID : {roomId}</RoomID>
-              <BackBtn isDark={isDark} onClick={handleBack}>
-                &times;
-              </BackBtn>
-            </ChatHeader>
+        <ChatContainer isDark={isDark}>
+          <ChatHeader isDark={isDark}>
+            <BackBtn isDark={isDark} onClick={handleBack}>
+              ◀ 이전으로
+            </BackBtn>
+            <RoomID isDark={isDark}>ID : {roomId}</RoomID>
+            <QuestionEndBtn
+              isDark={isDark}
+              onClick={() => {
+                patchChatEnd(accessToken.authorization, roomId);
+                stompClient.current.unsubscribe(`/sub/chat/room/${roomId}`);
+                handleBack();
+              }}
+            >
+              상담 종료하기
+            </QuestionEndBtn>
+          </ChatHeader>
 
-            <MsgLists isDark={isDark} ref={scrollRef}>
-              {ChatContents.data?.map((el: ChatContents) => {
-                const sendDate: string = el.sendDate.split(".")[0];
-                const sendTimes: string[] = sendDate.split("T")[1].split(":");
-                const sendTime =
-                  parseInt(sendTimes[0]) >= 12
-                    ? `오후 ${parseInt(sendTimes[0]) - 12}:${parseInt(
-                        sendTimes[1]
-                      )}`
-                    : `오전 ${parseInt(sendTimes[0])}:${parseInt(
-                        sendTimes[1]
-                      )}`;
-                if (el.sender === el.roomId) {
-                  return (
-                    <LeftMsgList>
-                      <MsgContent isDark={isDark}>{el.message}</MsgContent>
-                      <MsgDate isDark={isDark}>{sendTime}</MsgDate>
-                    </LeftMsgList>
-                  );
-                } else {
-                  return (
-                    <RightMsgList>
-                      <MsgDate isDark={isDark}>{sendTime}</MsgDate>
-                      <MsgContent isDark={isDark}>{el.message}</MsgContent>
-                    </RightMsgList>
-                  );
+          <MsgLists isDark={isDark} ref={scrollRef}>
+            {!isChatStart ? (
+              <ChatEmpty>
+                <ChatStart isDark={isDark} onClick={() => setChatStart(true)}>
+                  답변 시작하기
+                </ChatStart>
+              </ChatEmpty>
+            ) : (
+              <ChatList isArrive={isArrive} scrollRef={scrollRef} />
+            )}
+            <div ref={messageEndRef}></div>
+          </MsgLists>
+          <ChatFooter>
+            <ChatInputBox isDark={isDark}>
+              <ChatInput
+                isDark={isDark}
+                value={isMsg}
+                onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+                  setMsg(e.target.value)
                 }
-              })}
-              <div ref={messageEndRef}></div>
-            </MsgLists>
-            <ChatFooter>
-              <ChatInputBox isDark={isDark}>
-                <ChatInput
-                  isDark={isDark}
-                  value={isMsg}
-                  onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                    setMsg(e.target.value)
-                  }
-                  onKeyUp={(e: React.KeyboardEvent<HTMLInputElement>) =>
-                    handleKeyEnter(e)
-                  }
-                />
-                <ChatSendBtn isDark={isDark} onClick={handleSendMessage} />
-              </ChatInputBox>
-            </ChatFooter>
-          </ChatContainer>
-        )}
+                onKeyUp={(e: React.KeyboardEvent<HTMLInputElement>) =>
+                  handleKeyEnter(e)
+                }
+              />
+              <ChatSendBtn isDark={isDark} onClick={handleSendMessage} />
+            </ChatInputBox>
+          </ChatFooter>
+        </ChatContainer>
       </MainContainer>
     </PageContainer>
   );
@@ -267,7 +191,7 @@ export const ChatContainer = styled.div<DarkMode>`
   max-width: 800px;
   margin: 20px 0px;
   background-color: ${(props) =>
-    props.isDark ? globalTokens.Black.value : globalTokens.LightRed.value};
+    props.isDark ? globalTokens.Black.value : globalTokens.Background.value};
   border-radius: 8px;
 `;
 
@@ -277,7 +201,7 @@ export const ChatHeader = styled.div<DarkMode>`
   padding: 10px 10px;
   border-radius: 8px 8px 0px 0px;
   background-color: ${(props) =>
-    props.isDark ? globalTokens.Black.value : globalTokens.LightRed.value};
+    props.isDark ? globalTokens.Black.value : globalTokens.Background.value};
   border-bottom: 1px solid
     ${(props) =>
       props.isDark ? globalTokens.Gray.value : globalTokens.White.value};
@@ -293,22 +217,38 @@ export const RoomID = styled(BodyTextTypo)`
 
 export const BackBtn = styled(BodyTextTypo)`
   position: absolute;
-  right: 2%;
-  padding: 5px 10px;
+  left: 2%;
+  margin: 5px 5px;
   font-weight: ${globalTokens.Bold.value};
   cursor: pointer;
 `;
 
 export const QuestionEndBtn = styled(RegularButton)`
   position: absolute;
-  left: 2%;
+  right: 2%;
+  margin: 5px 5px;
   border: ${globalTokens.ThinHeight.value}px solid
     ${(props) =>
-      props.isDark ? globalTokens.Gray.value : globalTokens.Background.value};
+      props.isDark ? globalTokens.Gray.value : globalTokens.LightRed.value};
   background-color: ${(props) =>
-    props.isDark ? globalTokens.MainNavy.value : globalTokens.Background.value};
+    props.isDark ? globalTokens.MainNavy.value : globalTokens.LightRed.value};
   font-weight: 600;
   font-size: 14px;
+`;
+
+export const ChatEmpty = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`;
+
+export const ChatStart = styled(RegularButton)`
+  width: 50%;
+  background-color: ${(props) =>
+    props.isDark ? globalTokens.MainNavy.value : globalTokens.LightRed.value};
 `;
 
 export const MsgLists = styled.ul<DarkMode>`
@@ -317,7 +257,7 @@ export const MsgLists = styled.ul<DarkMode>`
   padding: 15px 10px;
   overflow-y: scroll;
   background-color: ${(props) =>
-    props.isDark ? globalTokens.Black.value : globalTokens.LightRed.value};
+    props.isDark ? globalTokens.Black.value : globalTokens.Background.value};
   display: flex;
   flex-direction: column;
 
@@ -332,47 +272,15 @@ export const MsgLists = styled.ul<DarkMode>`
     background: ${(props) =>
       props.isDark
         ? globalTokens.LightNavy.value
-        : globalTokens.Background.value};
+        : globalTokens.Negative.value};
     border-radius: 10px;
   }
 
   &::-webkit-scrollbar-track {
     /*스크롤바 뒷 배경 색상*/
     background: ${(props) =>
-      props.isDark ? globalTokens.Black.value : globalTokens.LightRed.value};
+      props.isDark ? globalTokens.Black.value : globalTokens.Background.value};
   }
-`;
-
-export const MsgList = styled.li`
-  width: 100%;
-  padding: 10px;
-  display: flex;
-  align-items: end;
-`;
-
-export const LeftMsgList = styled(MsgList)`
-  justify-content: start;
-`;
-
-export const RightMsgList = styled(MsgList)`
-  justify-content: end;
-`;
-
-export const MsgContent = styled(BodyTextTypo)`
-  padding: 8px 18px;
-  background-color: ${(props) =>
-    props.isDark ? "rgba(255,255,255,0.15)" : globalTokens.White.value};
-  border-radius: 8px;
-  font-weight: ${globalTokens.Bold.value};
-`;
-
-export const MsgDate = styled.div<DarkMode>`
-  margin: 0px 10px;
-  padding-bottom: 2px;
-  font-size: 12px;
-  font-weight: 600;
-  color: ${(props) =>
-    props.isDark ? globalTokens.LightGray.value : globalTokens.White.value};
 `;
 
 export const ChatFooter = styled.div`
@@ -390,9 +298,7 @@ export const ChatInputBox = styled.div<DarkMode>`
   justify-content: space-around;
   align-items: center;
   background-color: ${(props) =>
-    props.isDark
-      ? globalTokens.LightNavy.value
-      : globalTokens.Background.value};
+    props.isDark ? globalTokens.LightNavy.value : globalTokens.LightRed.value};
 `;
 
 export const ChatInput = styled.input<DarkMode>`
@@ -403,6 +309,7 @@ export const ChatInput = styled.input<DarkMode>`
   border: none;
   color: ${(props) =>
     props.isDark ? globalTokens.White.value : globalTokens.Black.value};
+  font-weight: 600;
 `;
 
 export const ChatSendBtn = styled(SendArrow)<DarkMode>`
